Test 1611

+++ lib/parser.hpp
@@ -37,7 +37,7 @@ class Parser {
         }
       }
       if(count%2==1){
-        temp_sample.x_.push_back(atol(*beg));
+        temp_sample.x_.push_back(stoi(*beg));
         // LOG(INFO) << *beg;
       }
       count++;
